+++ src/haddock/core/exceptions.py
