+++ src/haddock/libs/libsubprocess.py
@@ -4,7 +4,7 @@ import shlex
 import subprocess
 
 from haddock.core.defaults import CNS_EXE, NUM_CORES
-from haddock.error import CNSRunningError, JobRunningError
+from haddock.core.exceptions import CNSRunningError, JobRunningError
 from haddock.libs.libparallel import Scheduler
