+++ src/haddock/libs/libworkflow.py
@@ -3,7 +3,7 @@ import logging
 import importlib
 import shutil
 from pathlib import Path
-from haddock.error import HaddockError, StepError
+from haddock.core.exceptions import HaddockError, StepError
 # unused
 # from haddock.core.defaults import MODULE_PATH_NAME, TOPOLOGY_PATH
 from haddock.modules import modules_category
