+++ src/haddock/modules/__init__.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import toml
 
-from haddock.error import StepError
+from haddock.core.exceptions import StepError
 from haddock.libs.libontology import ModuleIO
 from haddock.core.defaults import MODULE_PATH_NAME, MODULE_IO_FILE, TOPOLOGY_PATH
