+++ src/haddock/modules/topology/topoaa/__init__.py
@@ -10,7 +10,7 @@ from haddock.cns.util import (
     prepare_single_input,
     )
 from haddock.core.defaults import TOPOLOGY_PATH
-from haddock.error import StepError
+from haddock.core.exceptions import StepError
 from haddock.libs.libparallel import Scheduler
 from haddock.libs.libpdb import PDBFactory
 from haddock.libs.libsubprocess import CNSJob
